requirements.txt
numpy==1.24.3
pandas==2.0.1
psycopg2_binary==2.9.6
python-dotenv==1.0.0
